==> 3.4 - Data Preparation.ipynb <==
Tratamentos para serem realizados:

1.   Selecionar apenas os atributos: 'TP_DEPENDENCIA_ADM_ESC', 'NU_NOTA_CN', 'NU_NOTA_CH', 'NU_NOTA_LC',
       'TP_LINGUA', 'NU_NOTA_COMP1', 'NU_NOTA_COMP2', 'NU_NOTA_COMP3',
       'NU_NOTA_COMP4', 'NU_NOTA_COMP5', 'NU_NOTA_REDACAO', 'NU_NOTA_MT'
2.   Atribuir valor 0 nos valores nulos
3.   Binarizar atributo 'TP_DEPENDENCIA_ADM_ESC'

from google.colab import drive
drive.mount('/content/drive')

### Carregar bibliotecas

import pandas as pd

### Carregar dados

train = pd.read_csv('drive/My Drive/Estudos/codenation/dados/train.csv', sep=',', index_col=0)

train.head()

test = pd.read_csv('drive/My Drive/Estudos/codenation/dados/test.csv', sep=',', index_col=None)

test.tail()

### Nos dados de treino, remover os atributos que não estão presentes nos dados de teste

test_columns = list(test.columns)
# Adicionando coluna alvo (y)
test_columns.append('NU_NOTA_MT')
train = train[test_columns]
train.head()

### Juntar arquivos de treino e teste para realizar os tratamentos necessários

train['tipo'] = 'train'
test['tipo'] = 'test'

dados = pd.concat([train, test])
dados.head()

### Selecionar apenas os atributos
'TP_DEPENDENCIA_ADM_ESC', 'NU_NOTA_CN', 'NU_NOTA_CH', 'NU_NOTA_LC',
       'TP_LINGUA', 'NU_NOTA_COMP1', 'NU_NOTA_COMP2', 'NU_NOTA_COMP3',
       'NU_NOTA_COMP4', 'NU_NOTA_COMP5', 'NU_NOTA_REDACAO', 'NU_NOTA_MT'

colunas_selecionadas = ['tipo', 'NU_INSCRICAO', 'TP_DEPENDENCIA_ADM_ESC', 'NU_NOTA_CN', 'NU_NOTA_CH', 'NU_NOTA_LC',
       'TP_LINGUA', 'NU_NOTA_COMP1', 'NU_NOTA_COMP2', 'NU_NOTA_COMP3', 'NU_NOTA_COMP4', 'NU_NOTA_COMP5', 'NU_NOTA_REDACAO', 'NU_NOTA_MT']

dados = dados[colunas_selecionadas]
dados.head()

### Atribuir valor 0 nos valores nulos

dados = dados.fillna(0)
dados.head()

### Binarizar atributo 'TP_DEPENDENCIA_ADM_ESC'

dados = pd.get_dummies(data=dados, columns=['TP_DEPENDENCIA_ADM_ESC'], drop_first=True)
dados.head()

### Separar os dados de treino e teste

tratado_train = dados[dados.tipo == 'train'].copy()
tratado_train.drop(columns=['tipo'], inplace=True)
tratado_train.head()

tratado_train.shape

tratado_test = dados[dados.tipo == 'test'].copy()
tratado_test.drop(columns=['tipo'], inplace=True)
tratado_test.head()

tratado_test.shape

Salvar os dados tratados em arquivo CSV

tratado_train.to_csv('dados/train_tratado-3_3.csv', sep=';', index=False)

tratado_test.to_csv('dados/test_tratado-3_3.csv', sep=';', index=False)

****
# Próxima etapa: "Modeling"
