==> 5 - Evaluation.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

### Carregar bibliotecas

import pandas as pd
import glob

### Lendo o arquivo com os resultados de validação




resultados_validacao = pd.read_csv('resultados_treino.csv', sep=';')

resultados_validacao.head(10)

##### Identificando as 15 melhores desempenhos com base na métrica 'mean_squared_log_error'

resultados_validacao.sort_values('mean_squared_log_error', ascending=True).head(15)

# Verificando a métrica 'mean_absolute_error'
resultados_validacao.sort_values('mean_absolute_error', ascending=True).head(15)

***

# Retreinar melhor modelo com os dados completos, sem split

from sklearn.model_selection import RandomizedSearchCV, GridSearchCV
from sklearn.preprocessing import RobustScaler
from sklearn.decomposition import FastICA
from sklearn.neighbors import KNeighborsRegressor
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.ensemble import GradientBoostingRegressor
import sklearn
import joblib
import ast

melhor_modelo = resultados_validacao.sort_values('mean_absolute_error', ascending=True).head(1)
melhor_modelo

pipeline = melhor_modelo['pipeline'].values[0]
pipeline = Pipeline([('standard_scaler', StandardScaler()), ('GradientBoostingRegressor', GradientBoostingRegressor())])
pipeline

#parametros = melhor_modelo['parametros'].values[0]
#parametros = ast.literal_eval(parametros)
#parametros
parametros = {
            'standard_scaler__with_mean': [True, False],
            'standard_scaler__with_std': [True, False],
            'GradientBoostingRegressor__loss': ('ls', 'lad', 'huber', 'quantile'),
            'GradientBoostingRegressor__criterion': ('friedman_mse', 'mse', 'mae'),
            'GradientBoostingRegressor__max_features': ('auto', 'sqrt', 'log2'),
            'GradientBoostingRegressor__n_estimators': [50, 100, 150],
            'GradientBoostingRegressor__warm_start': [True, False]
        }
parametros

cv = melhor_modelo['cross_validation_random_search'].values[0]
cv

random_state = melhor_modelo['random_state_random_search'].values[0]
random_state

# Ler dados de train completo
train = pd.read_csv('dados/train_tratado-3_2.csv', sep=';')
train.drop(columns=['NU_INSCRICAO'], inplace=True)
X = train.drop(columns=['NU_NOTA_MT'])
y = train['NU_NOTA_MT']

#modelo = GridSearchCV(estimator=pipeline, param_grid=parametros, cv=cv, n_jobs= -1, verbose=2)
modelo = RandomizedSearchCV(n_iter=10, estimator=pipeline, param_distributions=parametros, cv=5, random_state=1, n_jobs= -1, verbose=42)

modelo.fit(X, y)

nome_arquivo = 'modelos/melhor/GradientBoostingRegressor.pkl'
joblib.dump(modelo.best_estimator_, nome_arquivo)
