==> 6 - Deployment.ipynb <==
### Utilizar o melhor modelo para realizar a predição dos dados de test

from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import joblib
import glob
import numpy as np

dados = pd.read_csv('dados/test_tratado-3_2.csv', sep=';')
dados.head()

X = dados.drop(columns=['NU_INSCRICAO', 'NU_NOTA_MT'])
X.head()

caminho = glob.glob('modelos/melhor/Gr*')[0]
caminho

modelo = joblib.load(caminho)

predict = modelo.predict(X)
predict

dados['NU_NOTA_MT'] = np.around(predict, 1)
dados[['NU_INSCRICAO', 'NU_NOTA_MT']].head(10)

#caminho_answer = 'drive/My Drive/Estudos/codenation/submissões/answer-18.csv'
#dados[['NU_INSCRICAO', 'NU_NOTA_MT']].to_csv(caminho_answer, index=False)
